+++ test/test_skill.py
@@ -36,9 +36,6 @@ from ovos_utils.messagebus import FakeBus
 
 from mycroft.skills.skill_loader import SkillLoader
 from mycroft_bus_client import Message
-import datetime
-from neon_utils.skills.neon_skill import LOG
-import pytz
 
 import lingua_franca
 lingua_franca.load_language('en')
@@ -114,16 +111,6 @@ class TestSkill(unittest.TestCase):
                           {'context_key': 'MallParsing'})
         self.skill.user_request_handling(message)
 
-    
-    def test_en_open_stores_search(self):
-        store_info = [{'name': 'ABC Stores', 'hours': '9:30am – 9pm', 'location': 'Street Level 1, near Centerstage',
-                      'logo': 'https://gizmostorageprod.blob.core.windows.net/tenant-logos/1615937914061-abcstores.png'},
-                     {'name': 'ABC Stores', 'hours': '10:30am – 8:30pm', 'location': 'Street Level 1, in the Ewa Wing',
-                      'logo': 'https://gizmostorageprod.blob.core.windows.net/tenant-logos/1615937946329-abcstores.png'}]
-
-        now_hour, now_min = 10, 00
-        result_stores = self.skill.open_stores_search(store_info, now_hour, now_min)
-        self.assertEqual(result_stores[0], store_info[0])
 
     def test_en_time_calculation(self):
         store_info = [{'name': 'ABC Stores', 'hours': '9:30am – 9pm', 'location': 'Street Level 1, near Centerstage',
